processing comments
